Houses will now get a district assigned.
Added the Districts and theyre implementation. Reareanged the packages a little. SectorManager is now District Manager. The Class DistrictDecider creates  4 Districts of which 3 are fixed (type not position). There will be one of each District type and an additional fourth distirctz which cann be any of the other ones. The Position of the district is ranomized. Districts wrap Sector logic and theyre placeble logik can be used to invastigate for placing spots and collision detection.

The Logik which dictates that sectors cant overlapp is commented out, because it will need some math to calculate excakt sector boundaries, for the moment it is fine.